# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/auth/refresh":{"get":{"operationId":"AuthController_refreshTokens","parameters":[],"responses":{"201":{"description":"Successful"},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/auth/signup":{"post":{"operationId":"AuthController_onUserSignUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupDTO"}}}},"responses":{"200":{"description":"successful"},"201":{"description":"account successfully created"},"400":{"description":"bad request"},"401":{"description":"unauthorized user"}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_onUserLogin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"200":{"description":"successful"},"400":{"description":"bad request"},"409":{"description":"conflict responses"}},"tags":["Auth"]}},"/user/profile":{"get":{"operationId":"UsersController_findOneUser","summary":"","description":"Get single user","parameters":[],"responses":{"200":{"description":"successful"},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"Not found"}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/users":{"get":{"operationId":"UsersController_getAllUsers","summary":"","description":"Get single user","parameters":[],"responses":{"200":{"description":"successful"},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"Not found"}},"tags":["User"]}}},"info":{"title":"auth","description":"API Descriptions","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:9000"},{"url":"http://localhost:8080"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignupDTO":{"type":"object","properties":{"first_name":{"type":"string","example":"Oyin"},"last_name":{"type":"string","example":"Vamp"},"password":{"type":"string","example":"Vamp@123"},"email":{"type":"string","example":"vamp@gmail.com"},"date_of_birth":{"format":"date-time","type":"string","example":"2024-03-21 15:22:14.714376"},"phone_number":{"type":"string","example":"+2348036202050"}},"required":["first_name","last_name","password","email","date_of_birth","phone_number"]},"LoginDTO":{"type":"object","properties":{"email":{"type":"string","description":"user unique email","example":"vamp@gmail.com"},"password":{"type":"string","example":"Vamp@123"}},"required":["email","password"]}}}}